+++ src/ESPAsync_WiFiManager_Lite.h
@@ -9,7 +9,7 @@
   Built by Khoi Hoang https://github.com/khoih-prog/ESPAsync_WiFiManager_Lite
   Licensed under MIT license
   
-  Version: 1.8.1
+  Version: 1.8.2
    
   Version Modified By   Date        Comments
   ------- -----------  ----------   -----------
@@ -24,6 +24,7 @@
   1.7.0   K Hoang      09/01/2022  Fix the blocking issue in loop() with configurable WIFI_RECON_INTERVAL
   1.8.0   K Hoang      10/02/2022  Add support to new ESP32-S3
   1.8.1   K Hoang      11/02/2022  Add LittleFS support to ESP32-C3. Use core LittleFS instead of Lorol's LITTLEFS for v2.0.0+
+  1.8.2   K Hoang      21/02/2022  Optional Board_Name in Menu. Optimize code by using passing by reference
  *****************************************************************************************************************************/
 
 #pragma once
@@ -54,13 +55,13 @@
 #endif
 
 #ifndef ESP_ASYNC_WIFI_MANAGER_LITE_VERSION
-  #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION             "ESPAsync_WiFiManager_Lite v1.8.1"
+  #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION             "ESPAsync_WiFiManager_Lite v1.8.2"
   
   #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION_MAJOR       1
   #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION_MINOR       8
-  #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION_PATCH       1
+  #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION_PATCH       2
 
-  #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION_INT         1008001
+  #define ESP_ASYNC_WIFI_MANAGER_LITE_VERSION_INT         1008002
 #endif
 
 #ifdef ESP8266
@@ -358,8 +359,14 @@ typedef struct
 
 #define NUM_WIFI_CREDENTIALS      2
 
-// Configurable items besides fixed Header, just add board_name 
-#define NUM_CONFIGURABLE_ITEMS    ( ( 2 * NUM_WIFI_CREDENTIALS ) + 1 )
+#if USING_BOARD_NAME
+  // Configurable items besides fixed Header, just add board_name 
+  #define NUM_CONFIGURABLE_ITEMS    ( ( 2 * NUM_WIFI_CREDENTIALS ) + 1 )
+#else
+  // Configurable items besides fixed Header, just add board_name 
+  #define NUM_CONFIGURABLE_ITEMS    ( ( 2 * NUM_WIFI_CREDENTIALS ))
+#endif
+
 ////////////////
 
 #define HEADER_MAX_LEN            16
@@ -386,12 +393,20 @@ const char ESP_WM_LITE_HTML_HEAD_START[] /*PROGMEM*/ = "<!DOCTYPE html><html><he
 
 const char ESP_WM_LITE_HTML_HEAD_STYLE[] /*PROGMEM*/ = "<style>div,input{padding:5px;font-size:1em;}input{width:95%;}body{text-align: center;}button{background-color:#16A1E7;color:#fff;line-height:2.4rem;font-size:1.2rem;width:100%;}fieldset{border-radius:0.3rem;margin:0px;}</style>";
 
+#if USING_BOARD_NAME
 const char ESP_WM_LITE_HTML_HEAD_END[]   /*PROGMEM*/ = "</head><div style='text-align:left;display:inline-block;min-width:260px;'>\
 <fieldset><div><label>*WiFi SSID</label><div>[[input_id]]</div></div>\
 <div><label>*PWD (8+ chars)</label><input value='[[pw]]' id='pw'><div></div></div>\
 <div><label>*WiFi SSID1</label><div>[[input_id1]]</div></div>\
 <div><label>*PWD1 (8+ chars)</label><input value='[[pw1]]' id='pw1'><div></div></div></fieldset>\
 <fieldset><div><label>Board Name</label><input value='[[nm]]' id='nm'><div></div></div></fieldset>";	// DO NOT CHANGE THIS STRING EVER!!!!
+#else
+const char ESP_WM_LITE_HTML_HEAD_END[]   /*PROGMEM*/ = "</head><div style='text-align:left;display:inline-block;min-width:260px;'>\
+<fieldset><div><label>*WiFi SSID</label><div>[[input_id]]</div></div>\
+<div><label>*PWD (8+ chars)</label><input value='[[pw]]' id='pw'><div></div></div>\
+<div><label>*WiFi SSID1</label><div>[[input_id1]]</div></div>\
+<div><label>*PWD1 (8+ chars)</label><input value='[[pw1]]' id='pw1'><div></div></div></fieldset>";	// DO NOT CHANGE THIS STRING EVER!!!!
+#endif
 
 const char ESP_WM_LITE_HTML_INPUT_ID[]   /*PROGMEM*/ = "<input value='[[id]]' id='id'>";
 const char ESP_WM_LITE_HTML_INPUT_ID1[]  /*PROGMEM*/ = "<input value='[[id1]]' id='id1'>";
@@ -401,12 +416,21 @@ const char ESP_WM_LITE_FLDSET_START[]  /*PROGMEM*/ = "<fieldset>";
 const char ESP_WM_LITE_FLDSET_END[]    /*PROGMEM*/ = "</fieldset>";
 const char ESP_WM_LITE_HTML_PARAM[]    /*PROGMEM*/ = "<div><label>{b}</label><input value='[[{v}]]'id='{i}'><div></div></div>";
 const char ESP_WM_LITE_HTML_BUTTON[]   /*PROGMEM*/ = "<button onclick=\"sv()\">Save</button></div>";
+
+#if USING_BOARD_NAME
 const char ESP_WM_LITE_HTML_SCRIPT[]   /*PROGMEM*/ = "<script id=\"jsbin-javascript\">\
 function udVal(key,val){var request=new XMLHttpRequest();var url='/?key='+key+'&value='+encodeURIComponent(val);\
 request.open('GET',url,false);request.send(null);}\
 function sv(){udVal('id',document.getElementById('id').value);udVal('pw',document.getElementById('pw').value);\
 udVal('id1',document.getElementById('id1').value);udVal('pw1',document.getElementById('pw1').value);\
 udVal('nm',document.getElementById('nm').value);";
+#else
+const char ESP_WM_LITE_HTML_SCRIPT[]   /*PROGMEM*/ = "<script id=\"jsbin-javascript\">\
+function udVal(key,val){var request=new XMLHttpRequest();var url='/?key='+key+'&value='+encodeURIComponent(val);\
+request.open('GET',url,false);request.send(null);}\
+function sv(){udVal('id',document.getElementById('id').value);udVal('pw',document.getElementById('pw').value);\
+udVal('id1',document.getElementById('id1').value);udVal('pw1',document.getElementById('pw1').value);";
+#endif
 
 const char ESP_WM_LITE_HTML_SCRIPT_ITEM[]  /*PROGMEM*/ = "udVal('{d}',document.getElementById('{d}').value);";
 const char ESP_WM_LITE_HTML_SCRIPT_END[]   /*PROGMEM*/ = "alert('Updated');}</script>";
@@ -898,7 +922,7 @@ class ESPAsync_WiFiManager_Lite
     #define MIN_WIFI_CHANNEL      1
     #define MAX_WIFI_CHANNEL      11    // Channel 13 is flaky, because of bad number 13 ;-)
 
-    int setConfigPortalChannel(int channel = 1)
+    int setConfigPortalChannel(const int& channel = 1)
     {
       // If channel < MIN_WIFI_CHANNEL - 1 or channel > MAX_WIFI_CHANNEL => channel = 1
       // If channel == 0 => will use random channel from MIN_WIFI_CHANNEL to MAX_WIFI_CHANNEL
@@ -913,7 +937,8 @@ class ESPAsync_WiFiManager_Lite
     
     //////////////////////////////////////////////
     
-    void setSTAStaticIPConfig(const IPAddress& ip, const IPAddress& gw, const IPAddress& sn = IPAddress(255, 255, 255, 0),
+    void setSTAStaticIPConfig(const IPAddress& ip, const IPAddress& gw, 
+                              const IPAddress& sn = IPAddress(255, 255, 255, 0),
                               const IPAddress& dns_address_1 = IPAddress(0, 0, 0, 0),
                               const IPAddress& dns_address_2 = IPAddress(0, 0, 0, 0))
     {
@@ -936,7 +961,7 @@ class ESPAsync_WiFiManager_Lite
     
     //////////////////////////////////////////////
     
-    String getWiFiSSID(const uint8_t index)
+    String getWiFiSSID(const uint8_t& index)
     { 
       if (index >= NUM_WIFI_CREDENTIALS)
         return String("");
@@ -949,7 +974,7 @@ class ESPAsync_WiFiManager_Lite
     
     //////////////////////////////////////////////
 
-    String getWiFiPW(const uint8_t index)
+    String getWiFiPW(const uint8_t& index)
     {
       if (index >= NUM_WIFI_CREDENTIALS)
         return String("");
@@ -1022,6 +1047,13 @@ class ESPAsync_WiFiManager_Lite
     bool isConfigDataValid()
     {
       return hadConfigData;
+    }
+    
+    //////////////////////////////////////////////
+    
+    bool isConfigMode()
+    {
+      return configuration_mode;
     }
     
     //////////////////////////////////////////////
@@ -1332,7 +1364,7 @@ class ESPAsync_WiFiManager_Lite
 
     //////////////////////////////////////////////
     
-    void saveForcedCP(const uint32_t value)
+    void saveForcedCP(const uint32_t& value)
     {
       File file = FileFS.open(CONFIG_PORTAL_FILENAME, "w");
       
@@ -1368,7 +1400,7 @@ class ESPAsync_WiFiManager_Lite
     
     //////////////////////////////////////////////
     
-    void setForcedCP(const bool isPersistent)
+    void setForcedCP(const bool& isPersistent)
     {
       uint32_t readForcedConfigPortalFlag = isPersistent? FORCED_PERS_CONFIG_PORTAL_FLAG_DATA : FORCED_CONFIG_PORTAL_FLAG_DATA;
   
@@ -1966,7 +1998,7 @@ class ESPAsync_WiFiManager_Lite
 
     //////////////////////////////////////////////
     
-    void setForcedCP(const bool isPersistent)
+    void setForcedCP(const bool& isPersistent)
     {
       uint32_t readForcedConfigPortalFlag = isPersistent? FORCED_PERS_CONFIG_PORTAL_FLAG_DATA : FORCED_CONFIG_PORTAL_FLAG_DATA;
     
@@ -2539,7 +2571,10 @@ class ESPAsync_WiFiManager_Lite
             result.replace("[[pw]]",     ESP_WM_LITE_config.WiFi_Creds[0].wifi_pw);
             result.replace("[[id1]]",    ESP_WM_LITE_config.WiFi_Creds[1].wifi_ssid);
             result.replace("[[pw1]]",    ESP_WM_LITE_config.WiFi_Creds[1].wifi_pw);
+            
+#if USING_BOARD_NAME            
             result.replace("[[nm]]",     ESP_WM_LITE_config.board_name);
+#endif            
           }
           else
           {
@@ -2547,7 +2582,10 @@ class ESPAsync_WiFiManager_Lite
             result.replace("[[pw]]",  "");
             result.replace("[[id1]]", "");
             result.replace("[[pw1]]", "");
+            
+#if USING_BOARD_NAME            
             result.replace("[[nm]]",  "");
+#endif            
           }
           
 #if USE_DYNAMIC_PARAMETERS          
@@ -2622,7 +2660,10 @@ class ESPAsync_WiFiManager_Lite
         static bool pw_Updated  = false;
         static bool id1_Updated = false;
         static bool pw1_Updated = false;
+        
+#if USING_BOARD_NAME         
         static bool nm_Updated  = false;
+#endif 
           
         if (!id_Updated && (key == String("id")))
         {   
@@ -2672,18 +2713,19 @@ class ESPAsync_WiFiManager_Lite
           else
             strncpy(ESP_WM_LITE_config.WiFi_Creds[1].wifi_pw, value.c_str(), sizeof(ESP_WM_LITE_config.WiFi_Creds[1].wifi_pw) - 1);
         }
+#if USING_BOARD_NAME        
         else if (!nm_Updated && (key == String("nm")))
         {
           ESP_WML_LOGDEBUG(F("h:repl nm"));
           nm_Updated = true;
           
           number_items_Updated++;
-          
           if (strlen(value.c_str()) < sizeof(ESP_WM_LITE_config.board_name) - 1)
             strcpy(ESP_WM_LITE_config.board_name, value.c_str());
           else
             strncpy(ESP_WM_LITE_config.board_name, value.c_str(), sizeof(ESP_WM_LITE_config.board_name) - 1);
-        }
+        }    
+#endif
         
 #if USE_DYNAMIC_PARAMETERS
         else
@@ -2854,7 +2896,7 @@ class ESPAsync_WiFiManager_Lite
 
     //////////////////////////////////////////
 	
-	  void setMinimumSignalQuality(const int quality)
+	  void setMinimumSignalQuality(const int& quality)
 	  {
 	    _minimumQuality = quality;
 	  }
@@ -2862,7 +2904,7 @@ class ESPAsync_WiFiManager_Lite
 	  //////////////////////////////////////////
 
 	  //if this is true, remove duplicate Access Points - default true
-	  void setRemoveDuplicateAPs(const bool removeDuplicates)
+	  void setRemoveDuplicateAPs(const bool& removeDuplicates)
 	  {
 	    _removeDuplicateAPs = removeDuplicates;
 	  }
@@ -2982,7 +3024,7 @@ class ESPAsync_WiFiManager_Lite
 
 	  //////////////////////////////////////////
 
-	  int getRSSIasQuality(const int RSSI)
+	  int getRSSIasQuality(const int& RSSI)
 	  {
 	    int quality = 0;